+++ lib/extensions.dart
@@ -3,4 +3,7 @@ export 'package:olive/extensions/double_conversion_extensions.dart';
 export 'package:olive/extensions/int_constraint_extensions.dart';
 export 'package:olive/extensions/int_conversion_extensions.dart';
 export 'package:olive/extensions/iterable_filtering_extensions.dart';
+export 'package:olive/extensions/iterable_projection_extensions.dart';
+export 'package:olive/extensions/list_filtering_extensions.dart';
+export 'package:olive/extensions/stream_filtering_extensions.dart';
 export 'package:olive/extensions/string_logical_extensions.dart';