+++ lib/extensions/int_constraint_extensions.dart
@@ -1,22 +1,16 @@
 import 'dart:core';
 
 extension IntConversionExtensions on int {
-  /// <summary>
   /// Determines the minimum limitation of two values.
-  /// </summary>
   /// <param name="minValue">If this value is greater than {minValue}, this value is returned, owherwise, {minvalue} is returned.</param>
   int limitMin(int minValue) => this < minValue ? minValue : this;
 
-  /// <summary>
   /// Determines the maximum limitation of two values.
-  /// </summary>
   /// <param name="maxValue">If this value is smaller than {maxValue}, this value is returned, owherwise, {maxvalue} is returned.</param>
   int limitMax(int maxValue) => this > maxValue ? maxValue : this;
 
-  /// <summary>
   /// Determines the minimum and maximum limitation of two values.
   /// If this value is between {minValue} and {maxValue}, this value is returned. If this value is smaller than {minvalue}, {minvalue} is returned. If this value is greater than {maxvalue}, {maxvalue} is returned.
-  /// </summary>
   /// <param name="minValue">Determines the minimum value for comparing.</param>
   /// <param name="maxValue">Determines the maximum value for comparing.</param>
   int limitWithin(int minValue, int maxValue) =>
@@ -24,22 +18,16 @@ extension IntConversionExtensions on int {
 }
 
 extension NullableIntExtensions on int? {
-  /// <summary>
   /// Determines the minimum limitation of two values.
-  /// </summary>
   /// <param name="minValue">If this value is greater than {minValue}, this value is returned, owherwise, {minvalue} is returned.</param>
   int limitMin(int minValue) => (this ?? minValue).limitMin(minValue);
 
-  /// <summary>
   /// Determines the maximum limitation of two values.
-  /// </summary>
   /// <param name="maxValue">If this value is smaller than {maxValue}, this value is returned, owherwise, {maxvalue} is returned.</param>
   int limitMax(int maxValue) => (this ?? maxValue).limitMax(maxValue);
 
-  /// <summary>
   /// Determines the minimum and maximum limitation of two values.
   /// If this value is between {minValue} and {maxValue}, this value is returned. If this value is smaller than {minvalue}, {minvalue} is returned. If this value is greater than {maxvalue}, {maxvalue} is returned.
-  /// </summary>
   /// <param name="minValue">Determines the minimum value for comparing.</param>
   /// <param name="maxValue">Determines the maximum value for comparing.</param>
   int limitWithin(int minValue, int maxValue) =>